+++ miden-crypto/benches/smt-with-entries.rs
@@ -1,9 +1,10 @@
-use std::{fmt::Debug, hint, time::Duration};
+use std::{hint, time::Duration};
 
 use criterion::{BatchSize, BenchmarkId, Criterion, criterion_group, criterion_main};
-use miden_crypto::{Felt, ONE, Word, merkle::Smt};
-use rand_utils::prng_array;
-use winter_utils::Randomizable;
+use miden_crypto::merkle::Smt;
+
+mod common;
+use common::data::*;
 
 // 2^0, 2^4, 2^8, 2^12, 2^16
 const PAIR_COUNTS: [u64; 6] = [1, 16, 256, 4096, 65536, 1_048_576];
@@ -19,7 +20,7 @@ fn smt_with_entries(c: &mut Criterion) {
             b.iter_batched(
                 || {
                     // Setup.
-                    prepare_entries(pair_count, &mut seed)
+                    prepare_smt_entries(pair_count, &mut seed)
                 },
                 |entries| {
                     // Benchmarked function.
@@ -43,29 +44,4 @@ criterion_group! {
 criterion_main!(smt_with_entries_group);
 
 // HELPER FUNCTIONS
-// --------------------------------------------------------------------------------------------
-
-fn prepare_entries(pair_count: u64, seed: &mut [u8; 32]) -> Vec<(Word, Word)> {
-    let entries: Vec<(Word, Word)> = (0..pair_count)
-        .map(|i| {
-            let count = pair_count as f64;
-            let idx = ((i as f64 / count) * (count)) as u64;
-            let key = Word::new([generate_value(seed), ONE, Felt::new(i), Felt::new(idx)]);
-            let value = generate_word(seed);
-            (key, value)
-        })
-        .collect();
-    entries
-}
-
-fn generate_value<T: Copy + Debug + Randomizable>(seed: &mut [u8; 32]) -> T {
-    *seed = prng_array(*seed);
-    let value: [T; 1] = rand_utils::prng_array(*seed);
-    value[0]
-}
-
-fn generate_word(seed: &mut [u8; 32]) -> Word {
-    *seed = prng_array(*seed);
-    let nums: [u64; 4] = prng_array(*seed);
-    Word::new([Felt::new(nums[0]), Felt::new(nums[1]), Felt::new(nums[2]), Felt::new(nums[3])])
-}
+// === Helper Functions ===