+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix broken imports in CPU-specific `rescue` implementations (AVX2, SVE) ([#492](https://github.com/0xMiden/crypto/pull/492/)).
 - Added `{Smt,PartialSmt}::inner_node_indices` to make inner nodes accessible ([#494](https://github.com/0xMiden/crypto/pull/494)).
+- Added various benchmarks & related bench utilities ([#503](https://github.com/0xMiden/crypto/pull/503))
 
 ## 0.16.0 (2025-08-15)